It is all about the Love
And ... the hatred I have for Git.  I really feel like everyone just uses it because of Linus Torvalds, not because it is the "best solution".  Or else they can't be bothered to learn about Mercurial & have to take cheap shots at it.  Dude - just compare the commands.  Which makes more sense?  Which is easier to type from the command line?  SRSLY.